+++ backend2/src/main/java/com/cmoney/backend2/chat/service/ChatRoomService.kt
@@ -3,7 +3,6 @@ package com.cmoney.backend2.chat.service
 import com.cmoney.backend2.base.model.calladapter.RecordApi
 import com.cmoney.backend2.chat.service.api.chatroomsetting.response.ChatRoomSettingResponseBody
 import com.cmoney.backend2.chat.service.api.getallreport.request.ReportInfo
-import com.cmoney.backend2.chat.service.api.gethistorymessage.response.Message
 import com.cmoney.backend2.chat.service.api.getsubjectallrequests.response.AuthorizationRequestResponseBody
 import com.cmoney.backend2.chat.service.api.getuserprofile.response.UserProfileResponseBody
 import com.cmoney.backend2.chat.service.api.reportsomeone.request.ReportSomeone
@@ -217,7 +216,7 @@ interface ChatRoomService {
     suspend fun getMessageById(
         @Header("Authorization") authorization: String,
         @Path("messageId") messageId: Long
-    ): Response<Message>
+    ): Response<ResponseBody>
 
     /**
      * 取得歷史訊息, 新訊息優先
@@ -233,7 +232,7 @@ interface ChatRoomService {
         @Header("Authorization") authorization: String,
         @Path("chatroomId") roomId: Long,
         @QueryMap map: Map<String, String>
-    ): Response<List<Message?>>
+    ): Response<ResponseBody>
 
     /**
      * 取得歷史訊息, 舊訊息優先
@@ -249,7 +248,7 @@ interface ChatRoomService {
         @Header("Authorization") authorization: String,
         @Path("chatroomId") roomId: Long,
         @QueryMap map: Map<String, String>
-    ): Response<List<Message?>>
+    ): Response<ResponseBody>
 
     /**
      * 刪除訊息